Rename truffle-debug-utils to @truffle/debug-utils
