+++ packages/debug-utils/README.md
@@ -1,2 +1,2 @@
-# truffle-debug-utils
+# @truffle/debug-utils
 Code for integration between Truffle and Truffle Debugger
