+++ packages/truffle-core/lib/debug/interpreter.js
@@ -5,7 +5,7 @@ const path = require("path");
 const util = require("util");
 const ora = require("ora");
 
-const DebugUtils = require("truffle-debug-utils");
+const DebugUtils = require("@truffle/debug-utils");
 const selectors = require("truffle-debugger").selectors;
 const { session, solidity, trace, evm, controller } = selectors;
