+++ packages/truffle-core/lib/debug/printer.js
@@ -4,7 +4,7 @@ const debug = debugModule("lib:debug:printer");
 const path = require("path");
 const safeEval = require("safe-eval");
 
-const DebugUtils = require("truffle-debug-utils");
+const DebugUtils = require("@truffle/debug-utils");
 
 const selectors = require("truffle-debugger").selectors;
 const { session, solidity, trace, controller } = selectors;
